requirements.txt
python-dotenv==0.18.0
slackclient==2.9.3
slackeventsapi==2.2.1
